react: move app’s component lifecycle callback wrapper to compat.
