+++ src/compat/index.js
@@ -7,6 +7,7 @@ import extendStore from './extendStore';
 import createFactory from './createFactory';
 import createService from './createService';
 import makeMapToProps from './mapToProps';
+import makeGetMountableComponent from './getMountableComponent';
 
 export default {
   install(Frint) {
@@ -22,5 +23,6 @@ export default {
     Frint.createFactory = createFactory;
     Frint.createService = createService;
     Frint.mapToProps = makeMapToProps(Frint);
+    Frint.getMountableComponent = makeGetMountableComponent(Frint);
   }
 };
